+++ Sources/CryptoTickerApp/Services/Protocols/RefreshScheduling.swift
@@ -1,4 +1,7 @@
+import Foundation
+
+@MainActor
 protocol RefreshScheduling {
-    func start(_ action: @escaping () -> Void)
+    func start(interval: TimeInterval, _ action: @escaping () -> Void)
     func stop()
 }